app/requirements.txt
Django==3.1.1
gunicorn==20.0.4
psycopg2-binary==2.8.5
django-plotly-dash==1.4.2
channels-redis==3.1.0
django-redis==4.12.1
dpd-static-support==0.0.5
numpy==1.19.1